Fix sendMessage
Using the sendMessage, the text sent was always the first.
Debugging HTTP response, I realized that the connector was appending the query params.
The first time the call is going to be
/bot<token>/sendMessage?chat_id=1204698965&text=HelloHTTP/1.1
The second time the call is going to be
/bot<token>/sendMessage?chat_id=1204698965&chat_id=1204698965&text=Hello&text=ByeHTTP/1.1.

The problem was fixed by only instantiating a new HttpRequest instead of using the once that was instantiated at the beginning.